+++ polsarpro/io.py
@@ -38,160 +38,6 @@ from polsarpro.auxil import validate_dataset
 log = logging.getLogger(__name__)
 
 
-def read_T3_psp(input_dir: str):
-    """Reads a T3 matrix in the PolSARPro format.
-
-    Args:
-        input_dir (str): Input directory containing the elements of the T3 matrix and the configuration file config.txt.
-    """
-
-    input_path = Path(input_dir)
-
-    if not input_path.is_dir():
-        raise FileNotFoundError(f"Directory {input_dir} not found.")
-
-    pth_cfg = input_path / "config.txt"
-    if pth_cfg.is_file():
-        # read config and store in a dict
-        dict_cfg = {}
-        with open(pth_cfg, "r") as f:
-            lines = f.readlines()
-            lines = [line.replace("\n", "") for line in lines if "---" not in line]
-            for i in range(0, len(lines), 2):
-                dict_cfg[lines[i]] = lines[i + 1]
-    else:
-        raise FileNotFoundError("Configuration file config.txt does not exist.")
-
-    valid_mask_path = input_path / "mask_valid_pixels.bin"
-    is_valid_mask = True
-    if valid_mask_path.is_file():
-        valid_mask = read_psp_bin(valid_mask_path).astype(bool)
-    else:
-        is_valid_mask = False
-        log.info("Valid pixel mask not found, skipping.")
-
-    # image dimensions (azimuth, range)
-    naz = int(dict_cfg["Nrow"])
-    nrg = int(dict_cfg["Ncol"])
-
-    T3 = {}
-
-    for c in range(3):
-        for r in range(c + 1):
-            elt = f"T{r+1}{c+1}"
-            if r == c:
-                data = read_psp_bin(input_path / f"{elt}.bin", dtype="float32")
-
-                if is_valid_mask:
-                    data = np.where(valid_mask, data, np.nan)
-                T3[f"m{r+1}{c+1}"] = (("y", "x"), data)
-            else:
-                data_real = read_psp_bin(
-                    input_path / f"{elt}_real.bin", dtype="float32"
-                )
-                data_imag = read_psp_bin(
-                    input_path / f"{elt}_imag.bin", dtype="float32"
-                )
-                if is_valid_mask:
-                    data = np.where(valid_mask, data, np.nan + 1j * np.nan)
-                T3[f"m{r+1}{c+1}"] = (("y", "x"), data_real + 1j * data_imag)
-
-    attrs = {"poltype": "T3", "description": "Coherency matrix (3x3)"}
-    return xr.Dataset(T3, attrs=attrs).chunk({"x": 512, "y": 512})
-
-
-def read_C3_psp(input_dir: str):
-    """Reads a C3 matrix in the PolSARpro format.
-
-    Args:
-        input_dir (str): Input directory containing the elements of the C3 matrix and the configuration file config.txt.
-    """
-
-    input_path = Path(input_dir)
-
-    if not input_path.is_dir():
-        raise FileNotFoundError(f"Directory {input_dir} not found.")
-
-    pth_cfg = input_path / "config.txt"
-    if pth_cfg.is_file():
-        # read config and store in a dict
-        dict_cfg = {}
-        with open(pth_cfg, "r") as f:
-            lines = f.readlines()
-            lines = [line.replace("\n", "") for line in lines if "---" not in line]
-            for i in range(0, len(lines), 2):
-                dict_cfg[lines[i]] = lines[i + 1]
-    else:
-        raise FileNotFoundError("Configuration file config.txt does not exist.")
-
-    valid_mask_path = input_path / "mask_valid_pixels.bin"
-    is_valid_mask = True
-    if valid_mask_path.is_file():
-        valid_mask = read_psp_bin(valid_mask_path).astype(bool)
-    else:
-        is_valid_mask = False
-        log.info("Valid pixel mask not found, skipping.")
-
-    # image dimensions (azimuth, range)
-    naz = int(dict_cfg["Nrow"])
-    nrg = int(dict_cfg["Ncol"])
-
-    C3 = {}
-
-    for c in range(3):
-        for r in range(c + 1):
-            elt = f"C{r+1}{c+1}"
-            if r == c:
-                data = read_psp_bin(input_path / f"{elt}.bin", dtype="float32")
-
-                if is_valid_mask:
-                    data = np.where(valid_mask, data, np.nan)
-                C3[f"m{r+1}{c+1}"] = (("y", "x"), data)
-            else:
-                data_real = read_psp_bin(
-                    input_path / f"{elt}_real.bin", dtype="float32"
-                )
-                data_imag = read_psp_bin(
-                    input_path / f"{elt}_imag.bin", dtype="float32"
-                )
-                if is_valid_mask:
-                    data = np.where(valid_mask, data, np.nan + 1j * np.nan)
-                C3[f"m{r+1}{c+1}"] = (("y", "x"), data_real + 1j * data_imag)
-
-    attrs = {"poltype": "C3", "description": "Covariance matrix (3x3)"}
-    return xr.Dataset(C3, attrs=attrs).chunk("auto").chunk({"x": 512, "y": 512})
-
-
-def read_psp_bin(file_name: str, dtype: str = "float32"):
-    """Reads a raster bin file in the PolSARPro format.
-
-    Args:
-        file_name (str): Input file.
-    Note:
-        The parent directory must contain a config.txt file.
-    """
-    file_path = Path(file_name)
-    input_dir = file_path.parent
-
-    pth_cfg = input_dir / "config.txt"
-    if pth_cfg.exists():
-        # read config and store in a dict
-        cfg = {}
-        with open(pth_cfg, "r") as f:
-            lines = f.readlines()
-            lines = [line.replace("\n", "") for line in lines if "---" not in line]
-            for i in range(0, len(lines), 2):
-                cfg[lines[i]] = lines[i + 1]
-    else:
-        raise FileNotFoundError("Configuration file config.txt does not exist.")
-
-    # image dimensions (azimuth, range)
-    naz = int(cfg["Nrow"])
-    nrg = int(cfg["Ncol"])
-
-    return np.fromfile(file_path, dtype=dtype, count=naz * nrg).reshape((naz, nrg))
-
-
 def open_netcdf_beam(file_path: str | Path) -> xarray.Dataset:
     """Open data in the NetCDF-BEAM format exported by SNAP and create a valid python PolSARpro Dataset. Also works for complex matrix datasets written with polmat_to_netcdf.
 
@@ -390,7 +236,6 @@ def polmat_to_netcdf(ds: xarray.Dataset, file_path: str | Path):
                     data_out[f"{name}{i}{j}_real"] = arr.real
                     data_out[f"{name}{i}{j}_imag"] = arr.imag
 
-
     # make a new dataset with PolSARpro metadata
     # Preserve chunking when writing
     encoding = {var: {"chunksizes": data_out[var].data.chunksize} for var in data_out}
@@ -401,56 +246,3 @@ def polmat_to_netcdf(ds: xarray.Dataset, file_path: str | Path):
         coords=ds.coords,
     )
     ds_out.to_netcdf(file_path, encoding=encoding)
-
-# Reader below does not work due to a bug in gdal PALSAR driver (adds a line of zeros at the top)
-# def read_demo_data(data_dir: str) -> xarray.Dataset:
-#     """Reads ALOS-1 PALSAR San Francisco demo dataset.
-
-#     Args:
-#         data_dir (str): Input directory
-
-#     Returns:
-#         xarray.Dataset: Returns the elements of the Sinclair scattering matrix in a dataset.
-#     Note:
-#         The dataset must be downloaded at https://ietr-lab.univ-rennes1.fr/polsarpro-bio/sample_datasets/dataset/SAN_FRANCISCO_ALOS1.zip and unzipped in the directory of your choice.
-#     """
-
-#     # read metadata to extract the radiometric calibaration factor
-#     def extract_calibration_factor(data_path):
-#         from math import sqrt, pow
-
-#         file_path = data_path / "ceos_leader.txt"
-#         with open(file_path, "r") as file:
-#             for line in file:
-#                 if "Calibration Factor:" in line:
-#                     parts = line.strip().split()
-#                     # Assume the last element is the calibration value
-#                     try:
-#                         value = float(parts[-1])
-#                         # dB to linear
-#                         value = sqrt(pow(10.0, (value - 32.0) / 10.0))
-#                         return value
-#                     except ValueError("Invalid metadata."):
-#                         continue
-#         return None
-
-#     data_path = Path(data_dir)
-#     calfac = extract_calibration_factor(data_path=data_path)
-#     # silence warnings when dataset is in the SAR geometry
-#     with warnings.catch_warnings():
-#         warnings.filterwarnings("ignore", category=NotGeoreferencedWarning)
-#         # open with chunks to use dask arrays
-#         S = riox.open_rasterio(
-#             data_path / "VOL-ALPSRP202350750-P1.1__A",
-#             chunks="auto",
-#             # chunks=(-1, "auto", "auto"),
-#             band_as_variable=True,
-#         ).rename(band_1="hh", band_2="hv", band_3="vh", band_4="vv")
-#     # convert digital number to RCS
-#     S *= calfac
-
-#     # S['hv'] = 0.5*(S.hv + S.vh)
-#     # S['vh'] = S.hv
-#     # set polarimetric data type
-#     S.attrs = {"poltype": "S", "description": "Scattering matrix"}
-#     return S
